+++ docs/CHANGES.rst
@@ -7,6 +7,11 @@ Changelog
 - Added field IConfigurablePODTemplate.pod_formats to be able to select the
   format we want to generate the POD template in.
   [gbastien]
+- When evaluating the tal_condition on the template, pass extra_expr_ctx
+  to the TAL expression so 'context' becomes the element on which the TAL
+  expression is actually evaluated instead of the pod_template and 'template'
+  is the pod_template
+  [gbastien]
 
 
 0.1 (2015-07-17)