+++ docs/CHANGES.rst
@@ -6,6 +6,10 @@ Changelog
 - Renamed field 'pod_portal_type' to 'pod_portal_types' as this field
   is a multiselection field.
   [gbastien]
+- Renamed 'doc_uid' parameter used by the 'document-generation' view to
+  'template_uid', more obvious, and makes it available in the viewlet
+  link infos dict.
+  [gbastien]
 - Added field IConfigurablePODTemplate.pod_formats to be able to select the
   format we want to generate the POD template in.
   [gbastien]